speedtest-pwa/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
speedtest-cli>=2.1.3
